Add files via upload
Updated with the latest code version. Now meteors can come from any direction (even though they normally come from the right part of the screen). A new enemy was added, the turret. A particle system has been implemented such that we can make an object explode when it gets killed. Finally I have implemented ammo packages, which give the player ammo when caught, and which are tracked by the minimap.